requirements.txt
discord.py
googletrans==3.1.0a0
python-dotenv
language-tool-python